Remove supply_info.rs and move AssetSupply struct definition to issuance.rs
