+++ src/issuance.rs
@@ -20,8 +20,6 @@ use crate::note::{AssetBase, Nullifier, Rho};
 use crate::value::NoteValue;
 use crate::{Address, Note};
 
-use crate::supply_info::AssetSupply;
-
 /// Checks if a given note is a reference note.
 ///
 /// A reference note satisfies the following conditions:
@@ -29,6 +27,33 @@ use crate::supply_info::AssetSupply;
 /// - The note's recipient matches the reference recipient.
 fn is_reference_note(note: &Note) -> bool {
     note.value() == NoteValue::zero() && note.recipient() == ReferenceKeys::recipient()
+}
+
+/// Represents the amount of an asset, its finalization status and reference note.
+#[derive(Debug, Clone, Copy, Default)]
+#[cfg_attr(test, derive(PartialEq, Eq))]
+pub struct AssetSupply {
+    /// The amount of the asset.
+    pub amount: NoteValue,
+
+    /// Whether or not the asset is finalized.
+    pub is_finalized: bool,
+
+    /// The reference note, `None` if this `AssetSupply` instance is created from an issue bundle that does not include
+    /// a reference note (a non-first issuance)
+    pub reference_note: Option<Note>,
+}
+
+impl AssetSupply {
+    /// Creates a new AssetSupply instance with the given amount, finalization status and reference
+    /// note.
+    pub fn new(amount: NoteValue, is_finalized: bool, reference_note: Option<Note>) -> Self {
+        Self {
+            amount,
+            is_finalized,
+            reference_note,
+        }
+    }
 }
 
 /// A bundle of actions to be applied to the ledger.
